Articles and Books Managening Project
The System eliminates the chance of errors when performing its tasks, considerably reducing manual labor and enabling a smooth flow of searching and sorting activities. In this project, users would be able to sort, search and scrap the data (of the website). The end users of an Articles and Books managing project are usually people for whom libraries are intended (i.e. pupils, students, researchers, senior users, library staff, etc).